Create README.md
The creation of this website is for the purpose of putting into practice what has been learned throughout the various classes in the Full Stack Developer career. You can see through my remote repository on GitHub all the growth acquired as a developer, starting the practices as Front end Developer to soon perform the Back end. 
The main objective is to be able to show my capabilities that will help in the future either developing for startups or clients interested in bringing to the real world; through my skills what their imagination demands. Remember if you can believe it, you can create it.

Translated with www.DeepL.com/Translator (free version)